requirements.txt
blessed==1.19.0
click==8.0.1
importlib-metadata==4.8.3
inquirer==2.8.0
pkg-resources==0.0.0
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
six==1.16.0
typing-extensions==4.1.1
wcwidth==0.2.5
zipp==3.6.0
